Refactor theme context, move useTheme hook, and clean up commented code
