+++ src/components/SimpleDevToolsButton.jsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { useTheme } from '../contexts/ThemeContext';
+import { useTheme } from '../hooks/useTheme';
 
 /**
  * SimpleDevToolsButton Component
